requirements.txt
Flask==3.1.2
pandas==3.0.0
openpyxl==3.1.5
gunicorn==25.1.0
azure-storage-blob==12.24.0
